Applying changes to make MPI runtime usable with MSVC
- flyby: suppress integer conversion warnings
- flyby: suppress struct/class mismatch warnings

+++ flecsi/execution/mpi/runtime_driver.cc
@@ -33,7 +33,7 @@ remap_shared_entities()
 {
   // TODO: Is this superseded by index_map/reverse_index_map?
   auto& flecsi_context = context_t::instance();
-  const int my_color = flecsi_context.color();
+  const int my_color = static_cast<int>(flecsi_context.color());
 
   for (auto& coloring_info_pair : flecsi_context.coloring_info_map()) {
     auto index_space = coloring_info_pair.first;
@@ -57,7 +57,9 @@ remap_shared_entities()
     size_t index = 0;
     for (auto& shared : index_coloring.shared) {
       for (auto peer : shared.shared) {
-        MPI_Send(&index, 1, MPI_UNSIGNED_LONG_LONG, peer, 77, MPI_COMM_WORLD);
+        MPI_Send(
+            &index, 1, MPI_UNSIGNED_LONG_LONG, static_cast<int>(peer), 77,
+            MPI_COMM_WORLD);
       }
       new_shared.insert(
         flecsi::coloring::entity_info_t(shared.id, shared.rank, index, shared.shared));
@@ -71,7 +73,7 @@ remap_shared_entities()
 
     for (auto ghost : index_coloring.ghost) {
       MPI_Recv(&index, 1, MPI_UNSIGNED_LONG_LONG,
-               ghost.rank, 77, MPI_COMM_WORLD, &status);
+               static_cast<int>(ghost.rank), 77, MPI_COMM_WORLD, &status);
       new_ghost.insert(
         flecsi::coloring::entity_info_t(ghost.id, ghost.rank, index, {}));
     }
